updated controllers, error handling

+++ middleware/mutipleUpload.js
@@ -1,6 +1,6 @@
-const multer = require("multer");
-const fs = require("fs");
-const path = require("path");
+import multer from "multer";
+import fs from "fs";
+import path from "path";
 
 // Configure multer storage and file name
 const storage = multer.diskStorage({
@@ -23,40 +23,42 @@ const uploadMiddleware = (req, res, next) => {
       return res.status(400).json({ error: err.message });
     }
 
-    // Retrieve uploaded files
-    const files = req.files;
-    const errors = [];
+    if (typeof req.files !== "undefined") {
+      // Retrieve uploaded files
+      const files = req.files;
 
-    // Validate file types and sizes
-    files.forEach((file) => {
-      const allowedTypes = ["image/jpeg", "image/png"];
-      const maxSize = 5 * 1024 * 1024; // 5MB
+      const errors = [];
 
-      if (!allowedTypes.includes(file.mimetype)) {
-        errors.push(`Invalid file type: ${file.originalname}`);
-      }
-
-      if (file.size > maxSize) {
-        errors.push(`File too large: ${file.originalname}`);
-      }
-    });
-
-    // Handle validation errors
-    if (errors.length > 0) {
-      // Remove uploaded files
+      // Validate file types and sizes
       files.forEach((file) => {
-        fs.unlinkSync(file.path);
+        const allowedTypes = ["image/jpeg", "image/png"];
+        const maxSize = 5 * 1024 * 1024; // 5MB
+
+        if (!allowedTypes.includes(file.mimetype)) {
+          errors.push(`Invalid file type: ${file.originalname}`);
+        }
+
+        if (file.size > maxSize) {
+          errors.push(`File too large: ${file.originalname}`);
+        }
       });
 
-      return res.status(400).json({ errors });
+      // Handle validation errors
+      if (errors.length > 0) {
+        // Remove uploaded files
+        files.forEach((file) => {
+          fs.unlinkSync(file.path);
+        });
+
+        return res.status(400).json({ errors });
+      }
+
+      // Attach files to the request object
+      req.files = files;
     }
-
-    // Attach files to the request object
-    req.files = files;
-
     // Proceed to the next middleware or route handler
     next();
   });
 };
 
-module.exports = uploadMiddleware;
+export default uploadMiddleware;
